mapscraper_test_files also grabs and saves raw soup file from URL before cleaning it
